test: split keepout placement visual cases
